Move Mask component out of showcase and create an individual file for icons
