+++ src/components/icons/index.tsx
@@ -1,0 +1,1 @@
+export { GithubIcon } from './github'
